Revert "chore: merge current main into rollback notifications"
This reverts commit 715aab2e780963f01e8abc50325e0c1724015f94, reversing
changes made to e79cc040c95fc7d2e42e5242607eeaba38afe0b9.

+++ internal/test/conformance/state_manager_backend_test.go
@@ -295,7 +295,7 @@ func TestCommitteeMemberReadsPendingUpdateCommitteeProposal(t *testing.T) {
 	action, err := common.NewUpdateCommitteeGovAction(
 		nil,
 		nil,
-		map[*common.Credential]uint64{&coldCredential: 999},
+		map[*common.Credential]uint{&coldCredential: 999},
 		cbor.Rat{Rat: big.NewRat(2, 3)},
 	)
 	require.NoError(t, err)
